Abstract create and update job form components
